+++ CHANGES.rst
@@ -7,7 +7,10 @@ Version 1.1
 
 Unreleased
 
-* Support operating on multi-line records. (:issue:`1`)
+* Support operating on multi-line records (``linesieve --record-start``).
+  (:issue:`1`)
+* Allow parsing structured data (``linesieve parse``).
+  (:issue:`2`)
 * Allow piping lines to an arbitrary command (``linesieve pipe``).
 * Allow hiding sections (``linesieve hide``).